Refactor grpc
